Extract re-usable async infrastructure into 'core' module
